requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
requests==2.32.5
urllib3==2.5.0
python-dotenv==1.0.0
numpy>=1.24.0
pandas>=2.0.0
